+++ src/components/Perfil.js
@@ -14,8 +14,8 @@ class Perfil extends React.Component {
     const {isAuthenticated} = this.props.auth;
     const authLinks = (
       <div>
-        <h2>Perfil</h2>
-        Email del Usuario: <h4>{this.props.users.email}</h4>
+        <h2>Perfil de {this.props.user.name}</h2>
+        Email del Usuario: <h4>{this.props.user.email}</h4>
       </div>
     )
     const guestLinks = (
@@ -37,7 +37,7 @@ Perfil.propTypes = {
 function mapStateToProps(state, props) {
   return {
     auth: state.auth,
-    users: state.auth.users
+    user: state.auth.user
   }
 }